requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.9